requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appopener==1.7
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.2.0
confection==0.1.3
cymem==2.0.8
distro==1.8.0
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.18.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==1.3.7
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
preshed==3.0.9
psutil==5.9.6
PyAudio==0.2.14
PyAutoGUI==0.9.54
pydantic==2.4.2
pydantic-core==2.10.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
ttkbootstrap==1.10.1
typer==0.9.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.4
